Add an optional parameter to makeColorKeyTexture() to generate the old behaviour, i.e. zero the colour of keyed texels.  Add a comment in changes.txt explaining this, and update the tests.
This has revealed a problem in 16 bit textures in the software driver, in that it doesn't appear to respect the alpha of blended pixels: try example 06 with the software driver to see the problem.  "useAlphaChannelOfTexture" in draw2DImage() only worked in the software driver because the whole texel was colour 0. I'm investigating this now.

git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@2013 dfc29bdd-3216-0410-991c-e03cc46cb475