http://irrlicht.sourceforge.net/phpBB2/viewtopic.php?p=182554
Fix the EMT_TRANSPARENT_ALPHA_CHANNEL_REF material type in D3D9 and D3D8 by hard-coding the required alpha test value to 127 instead of taking it from the material.MaterialTypeParam.

Test case added, which also has a case for Burning; this is also broken, in a different way (it blends using the full alpha range, rather than a boolean check).  We can use this test to verify the fix for Burning when it happens.

git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@2055 dfc29bdd-3216-0410-991c-e03cc46cb475